feat(email-to-ticket): add 100% test coverage
